src/requirements.txt
pyaudio
aubio
numpy
rumps
noisereduce
torch
python-dotenv